Simplified loader for UW1/2 blocks. Reworked BuildObjectList and BuildTileMapUW to use uwblock loader Fixed layout of UW2 lev.ark save files to support saving of compressed/uncompressed source lev.ark files. Restored uw2pack usage in exporter tool Moved turning off of lightsources to a drop event rather than an update.

+++ UnityScripts/scripts/Conversations/ConversationVM.cs
@@ -273,7 +273,7 @@ public class ConversationVM : UWEBase {
 						{
 								if (isCompressed == 1)
 								{
-										int datalen=0;
+										long datalen=0;
 										char[] cnv_ark = DataLoader.unpackUW2(tmp_ark, add_ptr, ref datalen);
 										add_ptr=0;
 										/*
